+++ lib/src/backend/test_platform.dart
@@ -14,6 +14,10 @@ class TestPlatform {
   static const TestPlatform vm =
       const TestPlatform._("VM", "vm", isDartVm: true);
 
+  /// Dartium.
+  static const TestPlatform dartium = const TestPlatform._("Dartium", "dartium",
+      isBrowser: true, isBlink: true);
+
   /// Google Chrome.
   static const TestPlatform chrome = const TestPlatform._("Chrome", "chrome",
       isBrowser: true, isJS: true, isBlink: true);
@@ -23,7 +27,7 @@ class TestPlatform {
       isBrowser: true, isJS: true);
 
   /// A list of all instances of [TestPlatform].
-  static const List<TestPlatform> all = const [vm, chrome, firefox];
+  static const List<TestPlatform> all = const [vm, dartium, chrome, firefox];
 
   /// Finds a platform by its identifier string.
   ///