+++ test/runner/browser/runner_test.dart
@@ -24,6 +24,16 @@ void main() {
 }
 """;
 
+final _failure = """
+import 'dart:async';
+
+import 'package:test/test.dart';
+
+void main() {
+  test("failure", () => throw new TestFailure("oh no"));
+}
+""";
+
 void main() {
   setUp(() {
     _sandbox = createTempDir();
@@ -122,9 +132,23 @@ void main() {
       expect(result.exitCode, equals(0));
     });
 
+    test("on Dartium", () {
+      new File(p.join(_sandbox, "test.dart")).writeAsStringSync(_success);
+      var result = _runUnittest(["-p", "dartium", "test.dart"]);
+      expect(result.stdout, isNot(contains("Compiling")));
+      expect(result.exitCode, equals(0));
+    });
+
     test("on multiple browsers", () {
       new File(p.join(_sandbox, "test.dart")).writeAsStringSync(_success);
       var result = _runUnittest(["-p", "firefox", "-p", "chrome", "test.dart"]);
+      expect("Compiling".allMatches(result.stdout), hasLength(1));
+      expect(result.exitCode, equals(0));
+    });
+
+    test("on a JS and non-JS browser", () {
+      new File(p.join(_sandbox, "test.dart")).writeAsStringSync(_success);
+      var result = _runUnittest(["-p", "dartium", "-p", "chrome", "test.dart"]);
       expect("Compiling".allMatches(result.stdout), hasLength(1));
       expect(result.exitCode, equals(0));
     });
@@ -138,30 +162,20 @@ void main() {
 
   group("runs failing tests", () {
     test("on Chrome", () {
-      new File(p.join(_sandbox, "test.dart")).writeAsStringSync("""
-import 'dart:async';
-
-import 'package:test/test.dart';
-
-void main() {
-  test("failure", () => throw new TestFailure("oh no"));
-}
-""");
+      new File(p.join(_sandbox, "test.dart")).writeAsStringSync(_failure);
       var result = _runUnittest(["-p", "chrome", "test.dart"]);
       expect(result.exitCode, equals(1));
     });
 
     test("on Firefox", () {
-      new File(p.join(_sandbox, "test.dart")).writeAsStringSync("""
-import 'dart:async';
-
-import 'package:test/test.dart';
-
-void main() {
-  test("failure", () => throw new TestFailure("oh no"));
-}
-""");
+      new File(p.join(_sandbox, "test.dart")).writeAsStringSync(_failure);
       var result = _runUnittest(["-p", "firefox", "test.dart"]);
+      expect(result.exitCode, equals(1));
+    });
+
+    test("on Dartium", () {
+      new File(p.join(_sandbox, "test.dart")).writeAsStringSync(_failure);
+      var result = _runUnittest(["-p", "dartium", "test.dart"]);
       expect(result.exitCode, equals(1));
     });